pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Alexandr <alexandr.panteleev2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
django-rest-framework = "^0.1.0"
requests = "^2.31.0"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.2.0"
django-phonenumber-field = "^7.3.0"
django-phonenumbers = "^1.0.1"
django-filter = "^24.1"
drf-yasg = "^1.21.7"
setuptools = "^69.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
